some fixes

+++ telegram_bot.py
@@ -120,7 +120,8 @@ def download_yt_video(message):
             bot.reply_to(message, SERVER + '/videos/'+urllib.parse.quote(title_video) +'.mp4' )
             print( SERVER + '/videos/'+urllib.parse.quote(title_video) +'.mp4')
                 #REMOVER ARCHIVOS VIEJOS, SRT, MP3 y MP4
-            os.remove(title_video + '.mp4')
+            if tg_video_option=='sub':
+                os.remove(title_video + '.mp4')
             os.remove(title_video + '.srt')
             os.remove(title_video + '.mp3')
             tg_video = False
